use int instead of np.int in cython

+++ landlab/components/space_v2/cfuncs.pyx
@@ -9,7 +9,7 @@ from libc.math cimport log,exp
 DTYPE_FLOAT = np.double
 ctypedef np.double_t DTYPE_FLOAT_t
 
-DTYPE_INT = np.int
+DTYPE_INT = int
 ctypedef np.int_t DTYPE_INT_t
 
 DTYPE_UINT8 = np.uint8
